Removed NewtonsoftJson/SystemTextJson packages, determine serializer with preprocessors
