+++ GraphQL.Upload.AspNetCore.sln
@@ -13,10 +13,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{237BB3B8
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{CC98F604-7F80-4718-A307-7AA63FF55361}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Upload.AspNetCore.SystemTextJson", "src\GraphQL.Upload.AspNetCore.SystemTextJson\GraphQL.Upload.AspNetCore.SystemTextJson.csproj", "{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Upload.AspNetCore.NewtonsoftJson", "src\GraphQL.Upload.AspNetCore.NewtonsoftJson\GraphQL.Upload.AspNetCore.NewtonsoftJson.csproj", "{DB734621-019A-4656-BCFB-4DEBBD061568}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Upload.AspNetCore.NewtonsoftJson.Tests", "tests\GraphQL.Upload.AspNetCore.NewtonsoftJson.Tests\GraphQL.Upload.AspNetCore.NewtonsoftJson.Tests.csproj", "{9E1A348C-5299-42BB-9AE3-2E1B6195591E}"
 EndProject
 Global
@@ -37,14 +33,6 @@ Global
 		{2A45856C-FF34-47B8-BC5B-E037829ABCF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A45856C-FF34-47B8-BC5B-E037829ABCF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A45856C-FF34-47B8-BC5B-E037829ABCF9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DB734621-019A-4656-BCFB-4DEBBD061568}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DB734621-019A-4656-BCFB-4DEBBD061568}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DB734621-019A-4656-BCFB-4DEBBD061568}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DB734621-019A-4656-BCFB-4DEBBD061568}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9E1A348C-5299-42BB-9AE3-2E1B6195591E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9E1A348C-5299-42BB-9AE3-2E1B6195591E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9E1A348C-5299-42BB-9AE3-2E1B6195591E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -55,8 +43,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{41D23D02-FD11-457D-A8BC-1EC7726E85B2} = {237BB3B8-77C3-4190-91F1-5616A7507679}
-		{2E6132D3-535B-4AA9-BDFC-F33E5F925BE5} = {CC98F604-7F80-4718-A307-7AA63FF55361}
-		{DB734621-019A-4656-BCFB-4DEBBD061568} = {CC98F604-7F80-4718-A307-7AA63FF55361}
 		{9E1A348C-5299-42BB-9AE3-2E1B6195591E} = {237BB3B8-77C3-4190-91F1-5616A7507679}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
