--- src/GraphQL.Upload.AspNetCore.NewtonsoftJson/GraphQL.Upload.AspNetCore.NewtonsoftJson.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <AssemblyName>GraphQL.Upload.AspNetCore.NewtonsoftJson</AssemblyName>
-    <RootNamespace>GraphQL.Upload.AspNetCore.NewtonsoftJson</RootNamespace>
-    <PackageId>GraphQL.Upload.AspNetCore.NewtonsoftJson</PackageId>
-    <PackageTags>GraphQL Upload middleware AspNetCore Newtonsoft.Json</PackageTags>
-    <Description>GraphQL Upload middleware using Newtonsoft.Json for serialization</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GraphQL.Upload.AspNetCore\GraphQL.Upload.AspNetCore.csproj" />
-    <PackageReference Include="GraphQL.NewtonsoftJson" Version="$(GraphQLVersion)" />
-  </ItemGroup>
-
-</Project>
