--- src/GraphQL.Upload.AspNetCore.SystemTextJson/GraphQL.Upload.AspNetCore.SystemTextJson.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp3.1;netcoreapp3.0</TargetFrameworks>
-    <AssemblyName>GraphQL.Upload.AspNetCore.SystemTextJson</AssemblyName>
-    <RootNamespace>GraphQL.Upload.AspNetCore.SystemTextJson</RootNamespace>
-    <PackageId>GraphQL.Upload.AspNetCore.SystemTextJson</PackageId>
-    <PackageTags>GraphQL Upload middleware AspNetCore System.Text.Json</PackageTags>
-    <Description>GraphQL Upload middleware using System.Text.Json for serialization</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    <ProjectReference Include="..\GraphQL.Upload.AspNetCore\GraphQL.Upload.AspNetCore.csproj" />
-    <PackageReference Include="GraphQL.SystemTextJson" Version="$(GraphQLVersion)" />
-  </ItemGroup>
-
-</Project>
