+++ src/GraphQL.Upload.AspNetCore/GraphQL.Upload.AspNetCore.csproj
@@ -19,12 +19,18 @@
     <PackageReference Include="Microsoft.Extensions.Options" Version="$(MicrosoftExtensionsOptionsVersion)" />
   </ItemGroup>
 
+  <PropertyGroup Condition="'$(IsNetCore3OnwardsTarget)' == 'True'">
+    <DefineConstants>IS_NET_CORE_3_ONWARDS_TARGET</DefineConstants>
+  </PropertyGroup>
+
   <ItemGroup Condition="'$(IsNetCore3OnwardsTarget)' == 'True'">
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="GraphQL.SystemTextJson" Version="$(GraphQLVersion)" />
   </ItemGroup>
   
   <ItemGroup Condition="'$(IsNetCore3OnwardsTarget)' == 'False'">
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="$(MicrosoftAspNetCoreHttpAbstractionsVersion)" />
+    <PackageReference Include="GraphQL.NewtonsoftJson" Version="$(GraphQLVersion)" />
   </ItemGroup>
 
 </Project>
