--- src/GraphQL.Upload.AspNetCore/IGraphQLUploadRequestDeserializer.cs
@@ -1,9 +1,0 @@
-using Microsoft.AspNetCore.Http;
-
-namespace GraphQL.Upload.AspNetCore
-{
-    public interface IGraphQLUploadRequestDeserializer
-    {
-        GraphQLUploadRequestDeserializationResult DeserializeFromFormCollection(IFormCollection form);
-    }
-}
