+++ tests/GraphQL.Upload.AspNetCore.NewtonsoftJson.Tests/GraphQL.Upload.AspNetCore.NewtonsoftJson.Tests.csproj
@@ -16,7 +16,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\GraphQL.Upload.AspNetCore\GraphQL.Upload.AspNetCore.csproj" />
-    <ProjectReference Include="..\..\src\GraphQL.Upload.AspNetCore.NewtonsoftJson\GraphQL.Upload.AspNetCore.NewtonsoftJson.csproj" />
   </ItemGroup>
 
 </Project>
